feat: Save and fetch photos in Core Data, and download new photos
Any downloaded photos will now be saved in Core Data. App will check if
there are downloaded photos for a given pin so they do not need to be
redownloaded again but can be fetched from Core Data instead. When the
user taps on the "New Collection" button it will delete all currently
displayed photos and download a new set of photos from Flickr.